refactor: extract shared cloneOrPull into internal/gitops (#195)
Unify the duplicated git clone/pull + HTTP auth logic from
pkg/source/git and pkg/taskset into a single internal/gitops package.
The merged implementation preserves the more robust re-clone-on-corrupt
recovery from the taskset version. ~49 LOC net reduction.
